:hammer: Make GemtextText the base for more widgets
Also break up text.py.
