+++ src/rogallo/widgets/viewer/gemtext/heading.py
@@ -1,0 +1,49 @@
+"""Provides a widget for displaying a Gemtext heading."""
+
+##############################################################################
+# Gemtext imports.
+from gemtext import Heading, Line
+
+##############################################################################
+# Local imports.
+from .text import GemtextText
+
+
+##############################################################################
+class GemtextHeading(GemtextText):
+    """A widget for displaying a Gemtext heading."""
+
+    DEFAULT_CSS = """
+    GemtextHeading {
+        padding: 1 0;
+        &.--heading-1 {
+            color: $markdown-h1-color;
+            background: $markdown-h1-background;
+            text-style: $markdown-h1-text-style;
+            text-align: center;
+        }
+        &.--heading-2 {
+            color: $markdown-h2-color;
+            background: $markdown-h2-background;
+            text-style: $markdown-h2-text-style;
+        }
+        &.--heading-3 {
+            color: $markdown-h3-color;
+            background: $markdown-h3-background;
+            text-style: $markdown-h3-text-style;
+        }
+    }
+    """
+
+    def __init__(self, heading: Line) -> None:
+        """Initialize a Gemtext heading widget.
+
+        Args:
+            heading: The Gemtext heading to display.
+        """
+        super().__init__(heading)
+        assert isinstance(heading, Heading)
+        self.add_class(f"--heading-{heading.level}")
+
+
+### heading.py ends here
