+++ src/rogallo/widgets/viewer/gemtext/text.py
@@ -1,11 +1,13 @@
-"""Provides widgets that display plain text type content."""
+"""Provides the base widget for most forms of Gemtext text."""
 
 ##############################################################################
 # Gemtext imports.
-from gemtext import Heading, Line
+from gemtext import Line
+from rich.text import Text
 
 ##############################################################################
 # Textual imports.
+from textual.content import Content
 from textual.widgets import Static
 
 ##############################################################################
@@ -17,77 +19,19 @@ from .content_filter import GemtextContent
 class GemtextText(Static):
     """A widget for displaying a block of Gemtext text."""
 
-    def __init__(self, line: Line) -> None:
-        """Initialize a Gemtext widget.
+    def __init__(
+        self, line: Line | Content | Text | str, classes: str | None = None
+    ) -> None:
+        """Initialise the object.
 
         Args:
             line: The Gemtext line to display.
         """
-        super().__init__(GemtextContent.filter(line), markup=False)
-
-
-##############################################################################
-class GemtextParagraph(GemtextText):
-    """A widget for displaying a Gemtext paragraph."""
-
-    DEFAULT_CSS = """
-    GemtextParagraph {
-        padding: 0 2;
-    }
-    """
-
-
-##############################################################################
-class GemtextHeading(GemtextText):
-    """A widget for displaying a Gemtext heading."""
-
-    DEFAULT_CSS = """
-    GemtextHeading {
-        padding: 1 0;
-        &.--heading-1 {
-            color: $markdown-h1-color;
-            background: $markdown-h1-background;
-            text-style: $markdown-h1-text-style;
-            text-align: center;
-        }
-        &.--heading-2 {
-            color: $markdown-h2-color;
-            background: $markdown-h2-background;
-            text-style: $markdown-h2-text-style;
-        }
-        &.--heading-3 {
-            color: $markdown-h3-color;
-            background: $markdown-h3-background;
-            text-style: $markdown-h3-text-style;
-        }
-    }
-    """
-
-    def __init__(self, heading: Line) -> None:
-        """Initialize a Gemtext heading widget.
-
-        Args:
-            heading: The Gemtext heading to display.
-        """
-        super().__init__(heading)
-        assert isinstance(heading, Heading)
-        self.add_class(f"--heading-{heading.level}")
-
-
-##############################################################################
-class GemtextQuote(GemtextText):
-    """A widget for displaying a Gemtext quote."""
-
-    DEFAULT_CSS = """
-    GemtextQuote {
-        background: $boost;
-        border-left: outer $text-primary 50%;
-        padding: 0 1;
-        &:light {
-            border-left: outer $text-secondary;
-        }
-    }
-    """
+        super().__init__(
+            GemtextContent.filter(line) if isinstance(line, Line) else line,
+            markup=False,
+            classes=classes,
+        )
 
 
 ### text.py ends here
